improve files layout
