+++ web/src/App.tsx
@@ -23,11 +23,11 @@ declare global {
 }
 
 function App() {
-  const {  setBackEndUrl } = useStore(appStore)
+  const { setBackEndUrl } = useStore(appStore)
 
 
   useEffect(() => {
-   //  console.log("backedn url is", backend_url);
+    //  console.log("backedn url is", backend_url);
     // Properly handle async initialization
     const initBackend = async () => {
       const url = await initializeBackendUrl();
@@ -43,25 +43,27 @@ function App() {
       <MainLayout>
         <Routes>
           <Route path="" element={<Home />} />
-      <Route path="/files" element={<FilesPage />} />
-      <Route path="/files_all" element={<FilesPage />} />
-      <Route path="/files_workflow" element={<FilesPage />} />
-      <Route path="/files_templates" element={<FilesPage />} />
-      <Route path="/files_docs" element={<FilesPage />} />
-      <Route path="/files_attestation" element={<FilesPage />} />
-      <Route path="/files_info" element={<FilesPage />} />
-      <Route path="/files_settings" element={<FilesPage />} />
-          
-          
+          <Route path="/files" element={<FilesPage />} />
+          <Route path="/files_shared" element={<FilesPage />} />
+          <Route path="/files_workflows" element={<FilesPage />} />
+          <Route path="/files_templates" element={<FilesPage />} />
+          <Route path="/files_docs" element={<FilesPage />} />
+          <Route path="/files_attestation" element={<FilesPage />} />
+          <Route path="/files_info" element={<FilesPage />} />
+          <Route path="/files_settings" element={<FilesPage />} />
+          <Route path="/files_document_signature" element={<FilesPage />} />
+          <Route path="/files_domain_attestation" element={<FilesPage />} />
+
+
           <Route path="/loading" element={<Loading />} />
           <Route path="/share/:identifier" element={<SharePage />} />
-          <Route path="/aqua-forms" element={<AquaForms /> } />
-          <Route path="/pdf-signer" element={<PdfSigner  fileData={null}  setActiveStep={(_one)=>{
+          <Route path="/aqua-forms" element={<AquaForms />} />
+          <Route path="/pdf-signer" element={<PdfSigner fileData={null} setActiveStep={(_one) => {
 
-          }}/> } />
-          <Route path="/workflow" element={<WorkFlowPage /> } />
+          }} />} />
+          <Route path="/workflow" element={<WorkFlowPage />} />
           <Route path="/form-generator" element={<FormGenerator />} />
-          <Route path="/attestation_addresses" element={<AttestationAddresses /> } />
+          <Route path="/attestation_addresses" element={<AttestationAddresses />} />
         </Routes>
       </MainLayout>
     </BrowserRouter>
